server/requirements.txt
blinker==1.8.2
click==8.2.1
dnspython==2.7.0
Flask==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mongoengine==0.29.1
packaging==24.1
pymongo==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
Werkzeug==3.0.4
